Re-download updated entries
The current behavior is to not repeat a download if the file already
exists on disk. This commit adds an additional check -- if the `updated` field of an OPDS entry is more recent than the last
modified time of the file then it will re-download the file and emit an
addDocument event.